requirements.txt
Lektor>=3.1.3,<3.2
pyyaml>=5,<6
